+++ pkg/iota-pico-lib-nodejs.d.ts
@@ -351,17 +351,20 @@ export interface ILogger {
 export interface INetworkClient {
     /**
      * Get data asynchronously.
+     * @param data The data to send.
      * @param additionalPath An additional path append to the endpoint path.
      * @param additionalHeaders Extra headers to send with the request.
      * @returns Promise which resolves to the object returned or rejects with error.
      */
-    get(additionalPath?: string, additionalHeaders?: {
+    get(data: {
+        [key: string]: any;
+    }, additionalPath?: string, additionalHeaders?: {
         [header: string]: string;
     }): Promise<string>;
     /**
      * Post data asynchronously.
-     * @param additionalPath An additional path append to the endpoint path.
      * @param data The data to send.
+     * @param additionalPath An additional path append to the endpoint path.
      * @param additionalHeaders Extra headers to send with the request.
      * @returns Promise which resolves to the object returned or rejects with error.
      */
@@ -369,25 +372,16 @@ export interface INetworkClient {
         [header: string]: string;
     }): Promise<string>;
     /**
-     * Get data as JSON asynchronously.
-     * @typeparam U The generic type for the returned object.
-     * @param additionalPath An additional path append to the endpoint path.
-     * @param additionalHeaders Extra headers to send with the request.
-     * @returns Promise which resolves to the object returned or rejects with error.
-     */
-    getJson<U>(additionalPath?: string, additionalHeaders?: {
-        [header: string]: string;
-    }): Promise<U>;
-    /**
-     * Post data as JSON asynchronously.
+     * Request data as JSON asynchronously.
      * @typeparam T The generic type for the object to send.
      * @typeparam U The generic type for the returned object.
-     * @param data The data to send.
+     * @param data The data to send as the JSON body.
+     * @param method The method to send with the request.
      * @param additionalPath An additional path append to the endpoint path.
      * @param additionalHeaders Extra headers to send with the request.
      * @returns Promise which resolves to the object returned or rejects with error.
      */
-    postJson<T, U>(data: T, additionalPath?: string, additionalHeaders?: {
+    json<T, U>(data?: T, method?: NetworkMethod, additionalPath?: string, additionalHeaders?: {
         [header: string]: string;
     }): Promise<U>;
     /**
@@ -516,6 +510,11 @@ export interface ITimeService {
 }
 
 /**
+ * Represents the http request methods.
+ */
+export type NetworkMethod = "GET" | "POST" | "PUT" | "DELETE" | "PATCH";
+
+/**
  * Represents the protocols for communicating.
  */
 export type NetworkProtocol = "http" | "https";
@@ -612,6 +611,12 @@ export class NetworkEndPoint implements INetworkEndPoint {
      * @param rootPath The path to the endpoint.
      */
     constructor(protocol: NetworkProtocol, host: string, port: number, rootPath?: string);
+    /**
+     * Create a network endpoint by parsing a uri.
+     * @param uri The uri to parse.
+     * @returns The network endpoint.
+     */
+    static fromUri(uri: string): INetworkEndPoint;
     /**
      * The protocol to access the endpoint with.
      * @returns The protocol.
@@ -2352,7 +2357,19 @@ export class BusinessError extends CoreError {
     }, innerError?: Error);
 }
 
-export {};
+/**
+ * Helper class for address signing.
+ * Original https://github.com/iotaledger/iota.lib.js/blob/master/lib/crypto/signing/signing.js
+ */
+export class AddressHelper {
+    /**
+     * Create a checksum for the trits.
+     * @param trits The trits to create the checksum for.
+     * @param checksumLength The length of the checksum.
+     * @returns the checksum as trytes.
+     */
+    static createChecksum(trits: Int8Array, checksumLength: number): string;
+}
 
 /**
  * Helper class for signing bundles.
@@ -3043,11 +3060,14 @@ export class NetworkClient implements INetworkClient {
     constructor(networkEndPoint: INetworkEndPoint, logger?: ILogger, timeoutMs?: number, httpClientRequest?: (options: http.RequestOptions | https.RequestOptions | string | URL, callback?: (res: http.IncomingMessage) => void) => http.ClientRequest);
     /**
      * Get data asynchronously.
+     * @param data The data to send.
      * @param additionalPath An additional path append to the endpoint path.
      * @param additionalHeaders Extra headers to send with the request.
      * @returns Promise which resolves to the object returned or rejects with error.
      */
-    get(additionalPath?: string, additionalHeaders?: {
+    get(data: {
+        [key: string]: any;
+    }, additionalPath?: string, additionalHeaders?: {
         [header: string]: string;
     }): Promise<string>;
     /**
@@ -3061,25 +3081,16 @@ export class NetworkClient implements INetworkClient {
         [header: string]: string;
     }): Promise<string>;
     /**
-     * Get data as JSON asynchronously.
-     * @typeparam U The generic type for the returned object.
-     * @param additionalPath An additional path append to the endpoint path.
-     * @param additionalHeaders Extra headers to send with the request.
-     * @returns Promise which resolves to the object returned or rejects with error.
-     */
-    getJson<U>(additionalPath?: string, additionalHeaders?: {
-        [header: string]: string;
-    }): Promise<U>;
-    /**
-     * Post data as JSON asynchronously.
+     * Request data as JSON asynchronously.
      * @typeparam T The generic type for the object to send.
      * @typeparam U The generic type for the returned object.
-     * @param data The data to send.
+     * @param data The data to send as the JSON body.
+     * @param method The method to send with the request.
      * @param additionalPath An additional path append to the endpoint path.
      * @param additionalHeaders Extra headers to send with the request.
      * @returns Promise which resolves to the object returned or rejects with error.
      */
-    postJson<T, U>(data: T, additionalPath?: string, additionalHeaders?: {
+    json<T, U>(data?: T, method?: NetworkMethod, additionalPath?: string, additionalHeaders?: {
         [header: string]: string;
     }): Promise<U>;
     /**